+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ignite-ui-exo/react
 
+## 2.0.0
+
+### Minor Changes
+
+- Add component Toast
+
 ## 1.0.1
 
 ### Patch Changes